pyproject.toml
[tool.poetry]
name = "flamingo"
version = "0.1.0"
description = "A photo contest web application."
authors = ["Lucas Connors <lucas@revolutiontech.ca>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
Django = "~3.0"
psycopg2 = "^2.8.5"
Pillow = "^7.1.2"
python-memcached = "^1.59"
sorl-thumbnail = "^12.6.3"
whitenoise = "^5.0.1"
django-s3-storage = "^0.13.0"
python-dotenv = "^0.13.0"
django-configurations = { version = "^2.2", extras = ["database"] }

[tool.poetry.dev-dependencies]
coverage = "^5.1"
django-pigeon = "^0.3.0"
black = "^19.10b0"
pyupgrade = "^2.3"
isort = "^4.3.21"

[tool.isort]
# https://github.com/psf/black/blob/master/docs/compatible_configs.md#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.black]
line-length = 88
target_version = ["py38"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
